Fixed reseting of directory when the user hits cancel.

+++ source/ASEmbed.py
@@ -104,11 +104,11 @@ class LevelEditor:
         #Create a new window with given dimensions.
         aboutWindow = Toplevel()
         aboutWindow.title("About ASEmbed:")
-        aboutWindow.geometry("350x275+50+50")
+        aboutWindow.geometry("350x250+50+50")
         
         #Add the about info.
         msg = Message(aboutWindow, width=325, text="ASEmbed:\n\nVersion: 0.1.2\nDate: Dec 14, 2010\n\nASEmbed written by:\n\nAaron Ryan: asembed@gmail.com\nhttp://aaron.homedns.org\n\nProject Details:\nhttp://aaron.homedns.org/redmine/projects/asembed\
-\n\n\nExecutables built using pyinstaller.\n")
+\n")
         msg.pack()
 
         #Add a close button
@@ -134,9 +134,11 @@ class LevelEditor:
         
     
     def getDirectory(self):
-        self.directory = tkFileDialog.askdirectory()
-        self.directoryBox.delete(0, END)
-        self.directoryBox.insert(0, self.directory)
+        newDir = tkFileDialog.askdirectory()
+        if newDir:
+            self.directory = newDir
+            self.directoryBox.delete(0, END)
+            self.directoryBox.insert(0, self.directory)
         
         
 root = Tk()
